Operator short type conversion added and narrowing check

+++ CustomInt.h
@@ -58,7 +58,7 @@ struct CustomInt {
   // == operator overload
   bool operator==(const CustomInt &other) const;
 
-  explicit operator int() const;
+  explicit operator uint16_t() const;
 
   friend std::ostream &operator<<(std::ostream &stream,
                                   const CustomInt &customInt);
